Refactor FAQ page
