+++ app/components/header/header.js
@@ -124,16 +124,16 @@ class Header extends React.Component {
                       </li>
                     </Translation>
                     <li className='u-text-semi'>
-                     <IfLinkExists to='faq_merchants'>
-                       <Link to='faq_merchants' pointer='faq_merchants.link_title'
-                         id='track-nav-faq' className='u-padding-Vxs u-padding-Hm u-block' />
-                     </IfLinkExists>
+                      <IfLinkExists to='faq_merchants'>
+                        <Link to='faq_merchants' pointer='faq_merchants.link_title'
+                        id='track-nav-faq' className='u-padding-Vxs u-padding-Hm u-block' />
+                      </IfLinkExists>
                     </li>
                     <li className='u-text-semi'>
-                     <IfLinkExists to='faq_pro_dashboard'>
-                       <Link to='faq_pro_dashboard' pointer='faq_pro_dashboard.link_title'
-                         id='track-nav-faq-pro-dashboard' className='u-padding-Vxs u-padding-Hm u-block' />
-                     </IfLinkExists>
+                      <IfLinkExists to='faq_pro_dashboard'>
+                        <Link to='faq_pro_dashboard' pointer='faq_pro_dashboard.link_title'
+                        id='track-nav-faq-pro-dashboard' className='u-padding-Vxs u-padding-Hm u-block' />
+                      </IfLinkExists>
                     </li>
                     <hr className='u-margin-Vs' />
                     <li>
